// src/types/GithubUser.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Schema for a GitHub User, either an individual or organization.
 */
export interface GithubUser {
  /**
   * The unique identifier of the entity.
   */
  github_id: number;
  /**
   * The user's GitHub login name.
   */
  login: string;
  /**
   * The user's GitHub node ID.
   */
  node_id: string;
  /**
   * URL to the user's avatar image.
   */
  avatar_url: string;
  /**
   * The user's Gravatar ID.
   */
  gravatar_id?: string | null;
  /**
   * The user's GitHub API URL.
   */
  url?: string;
  /**
   * The user's GitHub profile URL.
   */
  html_url: string;
  /**
   * URL to the user's followers.
   */
  followers_url?: string;
  /**
   * URL to the user's following list.
   */
  following_url?: string;
  /**
   * URL to the user's gists.
   */
  gists_url?: string;
  /**
   * URL to the user's starred repositories.
   */
  starred_url?: string;
  /**
   * URL to the user's subscriptions.
   */
  subscriptions_url?: string;
  /**
   * URL to the user's organizations.
   */
  organizations_url?: string;
  /**
   * URL to the user's repositories.
   */
  repos_url?: string;
  /**
   * URL to the user's events.
   */
  events_url?: string;
  /**
   * URL to the user's received events.
   */
  received_events_url?: string;
  /**
   * The type of the user (e.g., 'User', 'Organization').
   */
  type: "User" | "Organization";
  /**
   * Whether the user is a site administrator.
   */
  site_admin: boolean;
  /**
   * The user's full name.
   */
  name?: string | null;
  /**
   * The user's company.
   */
  company?: string | null;
  /**
   * The user's blog URL.
   */
  blog?: string | null;
  /**
   * The user's location.
   */
  location?: string | null;
  /**
   * The user's email address.
   */
  email?: string | null;
  /**
   * Whether or not the user has marked themselves as for hire
   */
  hireable?: boolean | null;
  /**
   * The user's bio.
   */
  bio?: string | null;
  /**
   * The user's Twitter username.
   */
  twitter_username?: string | null;
  /**
   * The number of public repositories the user has.
   */
  public_repos?: number;
  /**
   * The number of public gists the user has.
   */
  public_gists?: number;
  /**
   * The number of followers the user has.
   */
  followers?: number;
  /**
   * The number of users the user is following.
   */
  following?: number;
  /**
   * The date and time the user's account was created.
   */
  created_at?: string;
  /**
   * The date and time the user's account was last updated.
   */
  updated_at?: string;
  /**
   * The number of contributions the user has made to a repository.
   */
  contributions?: number;
  [k: string]: unknown;
}